requirements.txt
pip~=22.0.3
pytz~=2021.3
platformdirs~=2.4.0
Django~=3.2.9
sqlparse~=0.4.2
asgiref~=3.4.1
PyYAML~=6.0
numpy~=1.21.4
flake8~=3.9.2
pyflakes~=2.3.1
setuptools~=60.8.0
djangorestframework~=3.12.4
requests~=2.26.0
PyMySQL~=1.0.2